+++ tree.php
@@ -1511,7 +1511,7 @@ function tree_edit($partial = false) {
 					divdata += "<ul><li id='" + id + "' data-jstree='" + jsdata + "'>" + text + '</li></ul>';
 				});
 
-				$(id).jstree('destroy').html(divdata);
+				$(id).jstree('destroy').html(DOMPurify.sanitize(divdata));
 			}
 
 			$(id)